Extract fixtures to separate files
