+++ solidity/test/bridge/Bridge.Wallets.test.ts
@@ -9,7 +9,7 @@ import type { Bridge, BridgeStub, IWalletRegistry } from "../../typechain"
 import { NO_MAIN_UTXO } from "../data/sweep"
 import { ecdsaWalletTestData } from "../data/ecdsa"
 import { constants, ecdsaDkgState, walletState } from "../fixtures"
-import bridgeFixture from "./bridge-fixture"
+import bridgeFixture from "../fixtures/bridge"
 
 chai.use(smock.matchers)
